Update patch set 2
Patch Set 2:

I got sshtunnel added to global-requirements, in preparation for an import from Fedora. crudini can't have the same fate, as it's license incomptatible (GPLv2). Since we only call it on the command line, I'm hoping by removing the version minimum we can pull it whatever comes with RDO (it doesn't seem to be in CentOS 8 by default).

Patch-set: 2